DOC: more sphinxext cleanup
